requirements.txt
streamlit
sentence-transformers
transformers
scikit-learn
numpy
google-generativeai
